tests/requirements.txt
psycopg2-binary==2.9.9
pytest==8.1.1
python-dotenv==1.0.1
boto3==1.34.90
